requirements.txt
Django==2.2.3
djangorestframework==3.10.1
psycopg2==2.8.3
python-dateutil==2.8.0
python-decouple==3.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
